+++ frb_example/dart_minimal/test/minimal_test.dart
@@ -16,9 +16,4 @@ Future<void> main() async {
     print('Action: Call rust (after)');
   });
   print('Action: Configure tests (end)');
-
-  test('custom serializer', () async {
-    expect(
-        await functionUsingTypeWithCustomSerializer(arg: 123456789), 123456789);
-  });
 }
