+++ frb_example/pure_dart/test/api/custom_ser_des_test.dart
@@ -1,0 +1,14 @@
+// FRB_INTERNAL_GENERATOR: {"forbiddenDuplicatorModes": ["sync", "rustAsync", "sse", "sync sse", "rustAsync sse"]}
+
+import 'package:frb_example_pure_dart/src/rust/api/custom_ser_des.dart';
+import 'package:frb_example_pure_dart/src/rust/frb_generated.dart';
+import 'package:test/test.dart';
+
+Future<void> main({bool skipRustLibInit = false}) async {
+  if (!skipRustLibInit) await RustLib.init();
+
+  test('custom serializer', () async {
+    expect(
+        await functionUsingTypeWithCustomSerializer(arg: 123456789), 123456789);
+  });
+}
